requirements.txt
amqp==2.5.2
asgiref==3.2.3
billiard==3.6.2.0
celery==4.4.0
certifi==2020.4.5.1
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
dj-database-url==0.5.0
Django==3.0.3
django-ckeditor==5.9.0
django-crispy-forms==1.8.1
django-js-asset==1.2.2
django-rest-swagger==2.2.0
djangorestframework==3.11.0
gunicorn==20.0.4
idna==2.9
importlib-metadata==1.5.0
itypes==1.2.0
Jinja2==2.11.2
kombu==4.6.7
MarkupSafe==1.1.1
openapi-codec==1.3.2
Pillow==7.0.0
pytz==2019.3
redis==3.5.2
requests==2.23.0
simplejson==3.17.0
sqlparse==0.3.0
uritemplate==3.0.1
urllib3==1.25.9
vine==1.3.0
whitenoise==5.1.0
zipp==2.2.0
psycopg2==2.8.5
django-heroku==0.3.1
